requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
Django==4.0.1
djangorestframework==3.13.1
greenlet==1.1.2
httpie==3.0.2
idna==3.3
msgpack==1.0.3
multidict==6.0.2
psycopg2==2.9.3
Pygments==2.11.2
pynvim==0.4.3
PySocks==1.7.1
python-decouple==3.6
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
